hero: initial basic device characteristics
This uses the new tree layout and scripts wrote by @jesec, but all the
code and inclusions related to the (currently unavailable) new board
support package sources and prebuilt blobs has been removed.

The original code was designed to include the slsi sources when they
were found available or using prebuilt binaries (not the proprietary
ones, but blobs built from source code with AOSP in mind), but we don't
have those either. The only way to get to a working tree now is to use
complementary slsi-cm and proprietary blobs where possible.

Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>